requirements.txt
# --- PyTorch specific index for CUDA 12.8 ---
--extra-index-url https://download.pytorch.org/whl/cu128

# --- AI & Model (Core) ---
torch==2.9.0+cu128
torchvision==0.24.0+cu128
torchaudio==2.9.0+cu128
accelerate==1.12.0
bitsandbytes==0.49.0
transformers==4.51.0
safetensors==0.7.0

# --- Backend & API ---
fastapi==0.128.0
uvicorn==0.40.0
python-multipart==0.0.21
pydantic==2.12.5
aiofiles==24.1.0

# --- UI & Data Visualization ---
gradio==6.4.0
pandas==2.3.3
pillow==12.0.0

# --- Utilities ---
requests==2.32.5
tqdm==4.67.1
